Added CSS v2 (read description -- it is important)
I have added a lot -- responsiveness (still in the early stages, and only applying to the home page), extensive CSS for the about page, and more CSS for the home page. In this file, fonts have been imported from Google Fonts and used on my website for the first time, and images have been styled. Grid is used for the first time, and another flexbox has been created.

This will be the last sole CSS file -- from now on I will have separate CSS files for my home page, about page, and gallery page, due to the length of CSS v2, and the interference of the CSS used in one page with the CSS used in another page (this has been suffered from twice).